+++ account-kit/signer/src/errors.ts
@@ -18,6 +18,8 @@ export class NotAuthenticatedError extends BaseError {
 export class OAuthProvidersError extends BaseError {
   override name = "OAuthProvidersError";
   constructor() {
-    super("OAuth providers not found", { docsPath: "/react/getting-started" });
+    super("OAuth providers not found", {
+      docsPath: "/react/getting-started",
+    });
   }
 }