+++ site/sidebar/shared.ts
@@ -1,11 +1,11 @@
 import { SidebarItem } from "vocs";
 import { coreSidebar } from "./core.js";
 import { infraSidebar } from "./infra.js";
+import { reactNativeSidebar } from "./react-native.js";
 import { reactGuides } from "./react.js";
 import { signerSidebar } from "./signer.js";
 import { smartContractsSidebar } from "./smart-contracts.js";
 import { thirdPartySidebar } from "./third-party.js";
-import { reactNativeSidebar } from "./react-native/index.js";
 
 export const concepts: SidebarItem = {
   text: "Concepts",
@@ -41,7 +41,6 @@ export const concepts: SidebarItem = {
 export const resources: SidebarItem = {
   text: "Resources",
   items: [
-    { text: "React Native", items: reactNativeSidebar, collapsed: true },
     {
       text: "Third Party",
       items: thirdPartySidebar,
@@ -55,12 +54,17 @@ export const resources: SidebarItem = {
 };
 
 export const guides: (
-  section?: "react" | "core" | "infra" | "signer" | "contracts"
+  section?: "react" | "core" | "infra" | "signer" | "contracts" | "react-native"
 ) => SidebarItem[] = (section) => [
   {
     text: "React",
     items: reactGuides,
     collapsed: section !== "react",
+  },
+  {
+    text: "React Native",
+    items: reactNativeSidebar,
+    collapsed: section !== "react-native",
   },
   {
     text: "Other JS Frameworks",
@@ -85,5 +89,5 @@ export const guides: (
 ];
 
 export const sharedSidebar: (
-  section?: "react" | "core" | "infra" | "signer" | "contracts"
+  section?: "react" | "core" | "infra" | "signer" | "contracts" | "react-native"
 ) => SidebarItem[] = (section) => [...guides(section), resources];